requirements.txt
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.2
djangorestframework-firebase==0.3.0
firebase-admin==2.18.0
gcloud==0.18.3
google-api-core==1.30.0
google-api-python-client==2.11.0
google-auth==1.32.1
google-auth-httplib2==0.1.0
google-cloud-core==1.7.1
google-cloud-firestore==2.1.3
google-cloud-storage==1.40.0
google-crc32c==1.1.2
google-resumable-media==1.3.1
googleapis-common-protos==1.53.0
Pyrebase==3.0.27
Pyrebase4==4.5.0
